Pass code theming options to RTEditor component - Delete MarkdownText.scss
